+++ CHANGELOG.md
@@ -10,7 +10,9 @@
     - spotbugsMaxRank setting allows to filter low-ranked rules (ranks are different from priorities!) (#15)
         Option modifies excludes.xml file (the only way to apply rank threshold)
     - spotbugsPlugin shortcut method to simplify plugins declaration (without afterEvaluate block or manual spotbugs plugin declaration)    
-* Spotbugs report now shows not only rule priority, but also it's rank: `[priority 2 / rank 14]`           
+* Spotbugs console report changes:
+    - Show both rule rank and priority : `[priority 2 / rank 14]`
+    - Identify rules from plugins: `[fb-contrib project | Correctness | FCBL_FIELD_COULD_BE_LOCAL]`            
 * Update codenarc 1.3 -> 1.4
 * Update checkstyle 8.17 -> 8.26
 * Update checkstyle config: